refactor: move config settings into the header as ConfigSettingsMenu
Rework the config-settings UI around the new page-header design.

- Remove the YAML-validation badge from the component (showValidationBadge
  prop, VALIDATION_BADGE, useYmlValidationStatus/useGetCiConfig). Where the
  validation status surfaces next is still open.
- Redesign it as a bare breadcrumb-style storage selector: a separator dot,
  a branch/globe icon, the source label (branch name in repo mode with an
  overflow tooltip + 240px truncation, or "Stored on Bitrise"), and a chevron
  control button that opens the existing menu (switch branch / download YAML /
  change storage). Only the trigger changed; the menu is unchanged.
- Rename ConfigSettingsBar -> ConfigSettingsMenu (it is no longer a bar) and
  move it into Header.tsx, right after the breadcrumb. The last breadcrumb
  crumb becomes "CI configuration".
- MainLayout: drop the CSS grid (only needed while the bar stacked above the
  nav) for a simple flex column. File tabs now render as a full-width strip
  directly under the header instead of beside the nav.
- Update Header / Navigation / ConfigSettingsMenu / YmlPage stories; swap
  MainLayout's legacy bitkit Box for the bitkit-v2 one.
