+++ source/javascripts/layouts/MainLayout.tsx
@@ -1,10 +1,9 @@
-import { Box } from '@bitrise/bitkit';
+import { Box } from '@chakra-ui/react/box';
 import { Redirect, Router, Switch } from 'wouter';
 
 import Header from '@/components/Header';
 import LazyRoute from '@/components/LazyRoute';
 import Navigation from '@/components/Navigation';
-import ConfigSettingsBar from '@/components/unified-editor/ConfigSettingsBar/ConfigSettingsBar';
 import RuntimeUtils from '@/core/utils/RuntimeUtils';
 import useHashLocation from '@/hooks/useHashLocation';
 import useHashSearch from '@/hooks/useHashSearch';
@@ -34,57 +33,27 @@ const MainLayout = () => {
   const isYmlPage = currentPath.startsWith(paths.yml);
   const isWebsiteMode = RuntimeUtils.isWebsiteMode();
 
-  // Modular configs get a global file-tab strip atop the editor area.
+  // Modular configs get a global file-tab strip directly under the header.
   const isModular = Boolean(useTree());
-  // With the bar present, the tab strip shares its 48px top row (so it doesn't jump
-  // between Visual and YAML mode); without it, the strip tops the content area instead.
-  const tabsBesideBar = isModular && isWebsiteMode;
+  // In website mode that strip spans the full width under the header (above the
+  // nav + content row); in CLI mode the tabs live inside the content column instead.
+  const tabsUnderHeader = isModular && isWebsiteMode;
 
   useMergedConfigSync();
 
-  // Key the YAML-page grid on tabsBesideBar (where the tab strip actually renders), not
-  // isModular: without the beside-bar strip the tabs live inside content, so the page is
-  // a single full-width column with no "tabs" area (matches the non-modular layout).
-  const gridTemplateColumns = isYmlPage && !tabsBesideBar ? '1fr' : '256px 1fr';
-  let gridTemplateAreas = tabsBesideBar ? `"bar tabs" "nav content"` : `"bar content" "nav content"`;
-  if (isYmlPage) {
-    gridTemplateAreas = tabsBesideBar ? `"bar tabs" "content content"` : `"bar" "content"`;
-  }
-
   return (
-    <Box h="100dvh" display="flex" flexDirection="column">
+    <Box height="100dvh" display="flex" flexDirection="column">
       <Header />
-      <Box
-        display="grid"
-        flex="1"
-        minH={0}
-        gridTemplateColumns={gridTemplateColumns}
-        gridTemplateRows="auto 1fr"
-        gridTemplateAreas={gridTemplateAreas}
-      >
-        {isWebsiteMode && (
-          <ConfigSettingsBar
-            gridArea="bar"
-            showValidationBadge={isYmlPage}
-            justifyContent={isYmlPage && !isModular ? 'flex-start' : 'space-between'}
-            borderRight={isYmlPage && !isModular ? undefined : '1px solid'}
-            borderRightColor={isYmlPage && !isModular ? undefined : 'border/regular'}
-          />
-        )}
-        {tabsBesideBar && (
-          <Box gridArea="tabs" minW={0}>
-            <OpenFileTabs />
-          </Box>
-        )}
+      {tabsUnderHeader && <OpenFileTabs />}
+      <Box display="flex" flex="1" minHeight={0}>
         {/* style instead of conditional unmount — keeps CI_CONFIG_RECEIVED / REQUEST_AI_DRAWER_OPEN listeners alive on the YAML page. */}
         <Navigation
-          gridArea="nav"
           borderRight="1px solid"
           borderColor="border/regular"
           style={{ display: isYmlPage ? 'none' : undefined }}
         />
-        <Box gridArea="content" display="flex" flexDirection="column" minW={0} minH={0}>
-          {isModular && !tabsBesideBar && <OpenFileTabs />}
+        <Box display="flex" flexDirection="column" flex="1" minWidth={0} minHeight={0}>
+          {isModular && !tabsUnderHeader && <OpenFileTabs />}
           <Box flex="1" overflowX="hidden" overflowY="auto">
             <Router hook={useHashLocation} searchHook={useHashSearch}>
               <InvalidYmlRedirect />
